accept patches from Georges - 10-configure.ac.patch removes a flag '-static' in a compilation - 20-manpages fixes a few typos in manpages (s/sp1/sp 1/) - 30-fixIntTypes fixes a more subtle bug, it make the definition of   integer types more precise and architecture-agnostic. See     http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=636927

+++ src/hfsplusclone.c
@@ -128,8 +128,8 @@ extern void readbitmap(char* device, image_head image_hdr, char* bitmap, int pui
     UInt32 bused = 0, bfree = 0, mused = 0;
     UInt32 block = 0, extent_block = 0, tb = 0, i = 0;
     int allocation_exten = 0;
-    long int allocation_start_block;
-    long int allocation_block_size;
+    UInt32 allocation_start_block;
+    UInt32 allocation_block_size;
     int start = 0;
     int bit_size = 1;
